+++ back/app/Http/Controllers/TaskLabelController.php
@@ -15,6 +15,7 @@ class TaskLabelController extends Controller
 
     public function store(Request $request, Project $project)
     {
+        $this->authorize('addLabel', $project);
         $validated = $request->validate(([
             'name' => ['required', 'string', 'max:50'],
             'color' => ['nullable', 'regex:/^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/'],
@@ -30,11 +31,13 @@ class TaskLabelController extends Controller
 
     public function show(TaskLabel $label)
     {
+        $this->authorize('view', $label);
         return response()->json($label);
     }
 
     public function update(Request $request, TaskLabel $label)
     {
+        $this->authorize('update', $label);
         $validated = $request->validate([
             'name' => ['sometimes', 'required', 'string', 'max:50'],
             'color' => ['nullable', 'regex:/^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/']
@@ -47,6 +50,7 @@ class TaskLabelController extends Controller
 
     public function destroy(TaskLabel $label)
     {
+        $this->authorize('delete', $label);
         $label->delete();
         return response()->json(null, 204);
     }
